refactor(execution): move folder execution under utils
