+++ src/commands/index.ts
@@ -1,2 +1,2 @@
 export * from './commit';
-export * from './execution';
+export * from '../utils/execution';
